+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unpublished yet]
 
 Highlights of the breaking changes:
-* A `.` is now required after a symbol interpretation in a theory or structure, e.g. `color := {red, blue, green}.`.
+* A `.` is now required after a symbol interpretation in a theory or structure, e.g. `color := {red, blue, green}.`. The string output of some API methods also has a `.` after an enumeration.
 
 ### IDP-Z3
 * (breaking) fix #182: Require a `.` after a symbol interpretation in a theory or structure